python-stt/requirements.txt
openai-whisper
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
